Comment resolutions
