+++ packages/smooth_app/lib/pages/onboarding/welcome_page.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:iso_countries/iso_countries.dart';
-import 'package:openfoodfacts/utils/CountryHelper.dart';
+import 'package:smooth_app/pages/onboarding/country_selector.dart';
+import 'package:smooth_app/pages/page_manager.dart';
 import 'package:smooth_app/themes/smooth_theme.dart';
 import 'package:smooth_ui_library/util/ui_helpers.dart';
 
@@ -13,11 +13,11 @@ class WelcomePage extends StatelessWidget {
   Widget build(BuildContext context) {
     final Size screenSize = MediaQuery.of(context).size;
     // Side padding is 8% of total width.
-    final double sidePadding = screenSize.width * 8 / 100;
+    final double sidePadding = screenSize.width * 0.08;
     // Top padding is 16% of total width.
-    final double topPadding = screenSize.height * 16 / 100;
+    final double topPadding = screenSize.height * 0.016;
     // Bottom padding is 4% of total width.
-    final double bottomPadding = screenSize.height * 4 / 100;
+    final double bottomPadding = screenSize.height * 0.04;
     final AppLocalizations appLocalizations = AppLocalizations.of(context)!;
     final TextStyle headlineStyle =
         Theme.of(context).textTheme.headline2!.apply(color: Colors.white);
@@ -72,7 +72,12 @@ class WelcomePage extends StatelessWidget {
                       primary: Colors.white,
                     ),
                     onPressed: () {
-                      // TODO(jasmeet): Navigate to the next page.
+                      // TODO(jasmeet): Check selected country and navigate to the next onboarding page.
+                      Navigator.push<Widget>(
+                          context,
+                          MaterialPageRoute<Widget>(
+                      builder: (BuildContext context) => PageManager(),)
+                      );
                     },
                     child: Text(
                       appLocalizations.next_label,
@@ -87,126 +92,5 @@ class WelcomePage extends StatelessWidget {
       ),
       backgroundColor: Theme.of(context).appBarTheme.backgroundColor,
     );
-  }
-}
-
-/// A selector for selecting user's country.
-class CountrySelector extends StatefulWidget {
-  const CountrySelector();
-
-  @override
-  State<CountrySelector> createState() => _CountrySelectorState();
-}
-
-class _CountrySelectorState extends State<CountrySelector> {
-  late List<Country> _countryList = <Country>[];
-  String? _chosenValue;
-  late Future<void> _initFuture;
-
-  @override
-  void initState() {
-    super.initState();
-    _initFuture = _init();
-  }
-
-  Future<void> _init() async {
-    final String locale = Localizations.localeOf(context).languageCode;
-    final List<Country> localizedCountries =
-        await IsoCountries.iso_countries_for_locale(locale);
-    _countryList = sanitizeCountriesList(localizedCountries);
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final AppLocalizations appLocalizations = AppLocalizations.of(context)!;
-    return FutureBuilder<void>(
-        future: _initFuture,
-        builder: (BuildContext context, AsyncSnapshot<void> snapshot) {
-          if (snapshot.hasError) {
-            return Text('Fatal Error: ${snapshot.error}');
-          }
-          if (snapshot.connectionState != ConnectionState.done) {
-            return const CircularProgressIndicator();
-          }
-          return Padding(
-            padding:
-                const EdgeInsets.only(top: MEDIUM_SPACE, bottom: LARGE_SPACE),
-            child: DropdownButtonFormField<String>(
-              value: _chosenValue,
-              style: const TextStyle(color: Colors.black),
-              decoration: InputDecoration(
-                enabledBorder: OutlineInputBorder(
-                  borderSide: const BorderSide(
-                      color: Color.fromARGB(255, 235, 235, 235)),
-                  borderRadius: BorderRadius.circular(VERY_LARGE_SPACE),
-                ),
-                filled: true,
-                fillColor: const Color.fromARGB(255, 235, 235, 235),
-              ),
-              items:
-                  _countryList.map<DropdownMenuItem<String>>((Country country) {
-                return DropdownMenuItem<String>(
-                  value: country.name,
-                  child: Text(country.name),
-                );
-              }).toList(),
-              hint: Text(
-                appLocalizations.country_chooser_label,
-                style: const TextStyle(
-                    color: Colors.black,
-                    fontSize: 16,
-                    fontWeight: FontWeight.w600),
-              ),
-              onChanged: (String? value) {
-                setState(() {
-                  if (value != null) {
-                    // TODO(jasmeet): Store pref in _userPreferences
-                    _chosenValue = value;
-                  }
-                });
-              },
-            ),
-          );
-        });
-  }
-
-  List<Country> sanitizeCountriesList(List<Country> localizedCountries) {
-    final List<Country> finalCountriesList = <Country>[];
-    final Map<String, OpenFoodFactsCountry> oFFIsoCodeToCountry =
-        <String, OpenFoodFactsCountry>{};
-    final Map<String, Country> localizedIsoCodeToCountry = <String, Country>{};
-    for (final OpenFoodFactsCountry c in OpenFoodFactsCountry.values) {
-      oFFIsoCodeToCountry.putIfAbsent(c.iso2Code.toLowerCase(), () => c);
-    }
-    for (final Country c in localizedCountries) {
-      localizedIsoCodeToCountry.putIfAbsent(
-          c.countryCode.toLowerCase(), () => c);
-    }
-    for (final String countryCode in oFFIsoCodeToCountry.keys) {
-      final Country? localizedCountry = localizedIsoCodeToCountry[countryCode];
-      if (localizedCountry == null) {
-        // No localization for the country name was found, use English name as
-        // default.
-        String countryName = oFFIsoCodeToCountry[countryCode]
-            .toString()
-            .replaceAll('OpenFoodFactsCountry.', '')
-            .replaceAll('_', ' ');
-        countryName =
-            '${countryName[0].toUpperCase()}${countryName.substring(1).toLowerCase()}';
-        final Country country =
-            Country(name: countryName, countryCode: countryName);
-        finalCountriesList.add(country);
-        continue;
-      }
-      // 'gb' is handled as 'uk' in the backend.
-      if (countryCode == 'gb') {
-        final Country modifiedCountry =
-            Country(name: localizedCountry.name, countryCode: 'uk');
-        finalCountriesList.add(modifiedCountry);
-        continue;
-      }
-      finalCountriesList.add(localizedCountry);
-    }
-    return finalCountriesList;
   }
 }
